refactor game logic types
