+++ game-logic/src/types/misc.ts
@@ -1,0 +1,3 @@
+import type { IGetUserStatsResult } from '@cards/db';
+
+export type UserStats = IGetUserStatsResult;
